Removed the call to fnameescape.
Vim's fnameescape function does not work well under Windows, because it
escapes spaces with a backslash, which are used as dir separators.
In windows most programs - including vim - are installed by default in
the "Program Files" folder - which breaks fnameescape.
Python handles paths better than Vim, and os.path.split does not need
the path to be escaped, so it is better left out.

I've also replaces os.path.split with os.path.dirname - which returns
only the dirname(instead of the [dirname,filename] array).